# openapi.yaml
openapi: 3.0.0
info:
  title: TrustIQ API
  description: "Complete user trust layer for your app. Assess any user in one API call — email risk, IP intelligence, device fingerprinting, behaviour signals, velocity checks, bot detection, and multi-account detection. Block fraud before it hits your platform."
  version: 1.0.0
servers:
  - url: https://api.trustiq.io/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Local development server
paths:
...
  /users/{user_id}/profile:
    get:
      tags:
        - User Trust Profiles
      summary: Get user trust profile
      description: Returns the complete trust profile built from all assessments and events tracked for this user.
      operationId: getUserProfile
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User's unique identifier
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{user_id}/linked-accounts:
    get:
      tags:
        - User Trust Profiles
      summary: Get linked accounts
      description: Returns all user_ids that share device fingerprints, IP addresses, or email domain patterns with this user.
      operationId: getLinkedAccounts
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User's unique identifier
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkedAccountsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /velocity/check:
    get:
      tags:
        - Velocity & Anomaly Intelligence
      summary: Check velocity for any entity
      description: Real-time velocity check against any entity — IP, email domain, or device.
      operationId: checkVelocity
      parameters:
        - name: entity_type
          in: query
          required: true
          schema:
            type: string
            enum: [ip, email_domain, device]
          description: Type of entity to check
        - name: entity_value
          in: query
          required: true
          schema:
            type: string
          description: Value of the entity to check
        - name: windows
          in: query
          required: false
          schema:
            type: string
            default: "1h,24h"
          description: Comma-separated list of time windows
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VelocityCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    UserProfileResponse:
      type: object
      required: [data]
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/UserProfileResponseData'
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/Meta'
...
    LinkedAccountsResponse:
      type: object
      required: [data]
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/LinkedAccountsResponseData'
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/Meta'
    VelocityCheckResponse:
      type: object
      required: [data]
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/VelocityCheckResponseData'
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/Meta'
...
    UserProfileResponseData:
      type: object
      properties:
        user_id:
          type: string
        trust_score:
          type: number
          minimum: 0
          maximum: 100
        risk_level:
          type: string
          enum: [critical, high, medium, low]
        account_age_days:
          type: integer
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        total_assessments:
          type: integer
        flagged_count:
          type: integer
        known_devices:
          type: array
          items:
            $ref: '#/components/schemas/KnownDevice'
        known_locations:
          type: array
          items:
            $ref: '#/components/schemas/KnownLocation'
        behaviour_patterns:
          $ref: '#/components/schemas/BehaviourPatterns'
        linked_accounts:
          type: array
          items:
            $ref: '#/components/schemas/LinkedAccount'
        risk_history:
          type: array
          items:
            $ref: '#/components/schemas/RiskHistory'
...
    LinkedAccountsResponseData:
      type: object
      properties:
        user_id:
          type: string
        linked_accounts:
          type: array
          items:
            $ref: '#/components/schemas/LinkedAccount'
        network_size:
          type: integer
        network_risk_level:
          type: string
    VelocityCheckResponseData:
      type: object
      properties:
        entity_type:
          type: string
        entity:
          type: string
        velocity:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/VelocityWindow'
        is_anomalous:
          type: boolean
        anomaly_score:
          type: number
        recommendation:
          type: string
...
    LinkedAccount:
      type: object
      properties:
        user_id:
          type: string
        link_type:
          type: string
          enum: [shared_device, shared_ip, shared_email_pattern]
        device_id:
          type: string
        ip:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        first_link_detected:
          type: string
        risk_implication:
          type: string
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        docs_url:
          type: string
    Meta:
      type: object
      properties:
        request_id:
          type: string
        version:
          type: string
        processing_ms:
          type: integer
        from_cache:
          type: boolean

    # Additional data schemas
...
    VelocityWindow:
      type: object
      properties:
        count:
          type: integer
        threshold:
          type: integer
        is_anomalous:
          type: boolean
...
    KnownDevice:
      type: object
      properties:
        device_id:
          type: string
        device_type:
          type: string
        browser:
          type: string
        first_seen:
          type: string
          format: date
        last_seen:
          type: string
          format: date
        is_primary:
          type: boolean
    KnownLocation:
      type: object
      properties:
        country:
          type: string
        city:
          type: string
        frequency_pct:
          type: number
    BehaviourPatterns:
      type: object
      properties:
        typical_login_hours:
          type: string
        avg_sessions_per_week:
          type: number
        avg_session_duration_mins:
          type: number
    RiskHistory:
      type: object
      properties:
        date:
          type: string
          format: date
        event:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 100
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    TooManyRequests:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-RapidAPI-Key